[BD-460] Use external services to resolve own pub IP.
This is done if the OpenStack meta service API endpoint
    http://169.254.169.254/2009-04-04/meta-data/public-ipv4
does not return a valid IP address. In testing found that
in the first 6.5+ minutes after booting, it returned an
empty string.  To solve this, use external services to
resolve the machine's own public IP.

Added a service list in SoftwareServerRestlet.conf: a space
separated URL list; added a dependent package in pom.xml.

Integration-tested with VM elasticity. Now once softwareserver is
up, job conversions between polyglot and it are working.