POL-166 Re-connect to RabbitMQ if the connection was closed/lost.
Tested. To test: use
    sudo service polyglot status
to find the pid, then
    sudo lsof -n -p <pid> | grep -i ipv | grep 141.142.227.65:amqp
to find the src port. Go to Rabbitmq -> connections, use "force close"
to close the connection. Then start a test. Without the fix, the
test will fail, and polyglot.log will have the exception messages;
with the fix, a line such as:
    [Mon Nov 16 10:32:47 2015] [steward]: RabbitMQ connection was closed,
submitting job failed. Re-created the connection.

will appear, RabbitMQ connection will be re-established, and the test
will succeed.